pyproject.toml
[project]
name = "tag"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.1",
    "pyserial>=3.5",
]
